style(layout): add responsive hook
